requirements.txt
requests==2.32.3
httpx==0.28.1
pytz==2025.1
tqdm==4.67.1
pyyaml==6.0.2
openai==1.65.2
pillow==11.1.0
python-dotenv==1.0.1
uvicorn==0.34.0
langchain==0.3.20
langchain-core==0.3.43
langchain-community==0.3.18
langchain-openai==0.3.7
langchain-chroma==0.2.2
langchain-ollama==0.2.3
langchain-experimental==0.3.4
langgraph-checkpoint-mongodb==0.1.1
langchain-taiga==1.0.2
langgraph==0.3.5
langgraph-cli[inmem]==0.1.74
unstructured==0.16.23
discord.py==2.5.0
duckduckgo-search==7.5.0
arxiv==2.1.3
spacy==3.8.4
wikipedia==1.4.0
youtube-search==2.1.2
playwright==1.50.0
python-taiga==1.3.2
pymongo==4.11.1
langsmith==0.3.11